Added binary decoder
